Make resuable rowAction hooks to be used in ig instances & instances list
